+++ ios/RCTConvert+BitmovinPlayer.swift
@@ -146,11 +146,6 @@ extension RCTConvert {
                 break
             }
         }
-#if !os(tvOS)
-        if let updatesNowPlayingInfoCenter = json["updatesNowPlayingInfoCenter"] as? Bool {
-            tweaksConfig.updatesNowPlayingInfoCenter = updatesNowPlayingInfoCenter
-        }
-#endif
         return tweaksConfig
     }